Fetch missed statuses
Get the last status id from Elasticsearch, crawl until we are up to date
again and only start streaming then. Also, if streaming is interrupted,
fetch missed statuses between retries.

This way, we shouldn't have any holes whenever crawling/streaming is
interrupted.

Also, we now try to go back to streaming if the stream is aborted and
don't go to crawling for the rest of time.