Refactor
The main change is to use this modules already existing data loading
method which does the same but is much shorter because we use existing
code.

Another noteworthy change is that now two files are output, one
containing the pure instance list, the other containing the whole
DataFrame, so we have that data available if we need it later.